+++ src/main/java/com/example/morpionsolitaire/models/Grid5T.java
@@ -16,35 +16,15 @@ import java.util.List;
 
 public class Grid5T extends Grid{
 
-    final static int WIDTH = 16;
-    final static int HEIGHT = 16;
-    final private String defaultGrid = "/src/main/java/com/example/morpionsolitaire/grids/default.grid";
-    private Cell[][] matrix = new Cell[WIDTH][WIDTH];
-    private List<Link> possibleMoves;
-    private boolean selectionInProcess = false;
-    private List<Link> movesHistory = new ArrayList<>();
-
     public Grid5T() throws IOException {
         this.load();
     }
-
-    /*private void load() throws IOException {
-        Path currentRelativePath = Paths.get("");
-        String s = currentRelativePath.toAbsolutePath().toString() + defaultGrid;
-        for (int i = 0 ; i< HEIGHT; i++) {
-            String read = Files.readAllLines(Paths.get(s)).get(i);
-            String[] line = read.split("");
-            for (int j = 0; j < WIDTH; j++){
-                matrix[i][j] = new Cell(i, j, Integer.parseInt(line[j]));
-            }
-        }
-    }*/
 
     public void setCell(int line, int column, int value){
         matrix[line][column].setValue(value);
     }
 
-    public Cell getCell(int i, int j){
+    public static Cell getCell(int i, int j){
         return matrix[i][j];
     }
 
@@ -380,22 +360,23 @@ public class Grid5T extends Grid{
 
     public void undoLastMove(){
         int index = this.movesHistory.size() - 1;
-        Link linkToRemove = this.movesHistory.get(index);
-        for (Cell c : linkToRemove.getNodes()){
-            c.unlink(linkToRemove.getLinkType());
+        if (index >= 0) {
+            Link linkToRemove = this.movesHistory.get(index);
+            for (Cell c : linkToRemove.getNodes()) {
+                c.unlink(linkToRemove.getLinkType());
+            }
+            Cell rootCell = linkToRemove.getRoot();
+            rootCell.setMainLink(LinkType.NONE);
+            this.matrix[rootCell.getI()][rootCell.getJ()].setValue(0);
+            rootCell.setLink(null);
+            movesHistory.remove(index);
         }
-        Cell rootCell = linkToRemove.getRoot();
-        rootCell.setMainLink(LinkType.NONE);
-        this.matrix[rootCell.getI()][rootCell.getJ()].setValue(0);
-        rootCell.setLink(null);
-        movesHistory.remove(index);
-
     }
 
 
-    public List<Link> getMovesHistory(){
+    /*public List<Link> getMovesHistory(){
         return movesHistory;
-    }
+    }*/
 }
